--- src/App/App.css
@@ -1,3 +1,0 @@
-.app {
-    height: 100vh;
-}
