+++ src/App/App.scss
@@ -1,0 +1,5 @@
+@import "src/styles/variables";
+
+.app {
+    height: 100vh;
+}
