+++ src/App/pages/TokensPage/SummaryCash/SummaryCash.js
@@ -1,15 +1,14 @@
 import * as React from 'react';
 import { Typography } from 'antd';
-import './SummaryCash.css';
+import './SummaryCash.scss';
+const { Title, Text } = Typography;
 
-const { Title } = Typography;
-const { Text } = Typography;
 
 const SummaryCash = ({totalSum}) => {
   return (
-    <div className={'monitor'}>
-      <Title id={'monitor__summary'} level={2}>${totalSum}</Title>
-      <Text id={'monitor__pubkey'} type="secondary">readonly</Text>
+    <div className="monitor">
+      <Title id="monitor__summary" level={2}>${totalSum}</Title>
+      <Text id="monitor__pubkey" type="secondary">readonly</Text>
     </div>
   )
 }
