+++ src/App/pages/TokensPage/Tokens/Rate/Rate.scss
@@ -1,0 +1,20 @@
+@import "src/styles/variables";
+
+
+.stonks {
+    font-size: 2.5rem;
+    &_up {
+        color: $color-green;
+
+    }
+    &_down {
+        color: $color-red;
+    }
+    &_static{
+        color: $color-light-dark;
+    }
+}
+
+.dollars {
+    font-size: 2.8rem;
+}
