+++ src/App/pages/TokensPage/Tokens/Rate/index.js
@@ -1,0 +1,1 @@
+export {default} from './Rate';
