+++ src/App/pages/TokensPage/Tokens/SingleToken/index.js
@@ -1,0 +1,1 @@
+export {default} from './SingleToken';
