--- src/App/pages/TokensPage/Tokens/Tokens.css
@@ -1,42 +1,0 @@
-.tokens-list {
-    background: white;
-    /*height: 100vh;*/
-}
-
-.ant-list-item {
-    padding-left: 16px;
-    padding-right: 16px;
-}
-
-.tokens-money-cont {
-    display: block;
-    text-align: right;
-}
-
-.tokens-money-cont__dollar{
-    color: rgba(0, 0, 0, 0.45);
-    font-size: 2.8rem;
-    line-height: 1.5715;
-}
-
-.stonks {
-    color: green;
-    font-size: 2.5rem;
-}
-
-.not-stonks{
-    color: red;
-    font-size: 2.5rem;
-}
-
-.ant-list-item-meta-title {
-    font-size: 3rem;
-}
-
-.tokens-money-cont__crypt{
-    font-size: 2.8rem;
-}
-
-.tokens-one-dollar{
-    font-size: 2.8rem;
-}
