+++ src/App/pages/TokensPage/Tokens/Tokens.scss
@@ -1,0 +1,16 @@
+@import "src/styles/variables";
+
+.tokens-list {
+    background: $color-white;
+}
+
+.ant-list-item {
+    padding: 2rem 3.6rem;
+}
+
+
+
+.ant-list-item-meta-title {
+    font-size: 3rem;
+}
+
