+++ src/App/pages/TokensPage/TokensPage.js
@@ -5,6 +5,7 @@ import Tokens from "./Tokens";
 import {someTokens} from "../../../utils/constants";
 import {formatMoney} from "../../../utils/formatMoney";
 
+
 const CreateTokens = () => {
   let totalSum = 0;
   const tokensForRender = [];
@@ -13,21 +14,21 @@ const CreateTokens = () => {
     symbol: 'ETH',
     name: 'Ethereum',
     balance: someTokens.ETH.balance,
-    price: someTokens.ETH.balance*someTokens.ETH.price.rate,
+    price: someTokens.ETH.balance * someTokens.ETH.price.rate,
     logo: '',
     rate: someTokens.ETH.price.rate,
     dif: someTokens.ETH.price.diff,
   });
   totalSum += tokensForRender[0].price;
   if (someTokens.tokens) {
-    someTokens.tokens.forEach((item)=>{
+    someTokens.tokens.forEach((item) => {
       if (item.tokenInfo.price !== false) {
         const itemInfo = item.tokenInfo;
-        const itemBalance = item.balance/Math.pow(10, itemInfo.decimals);
+        const itemBalance = item.balance / Math.pow(10, itemInfo.decimals);
 
-        const tmpPrice = itemBalance*itemInfo.price.rate;
+        const tmpPrice = itemBalance * itemInfo.price.rate;
 
-        totalSum+=tmpPrice;
+        totalSum += tmpPrice;
 
         tokensForRender.push({
           address: itemInfo.address,
@@ -35,7 +36,7 @@ const CreateTokens = () => {
           name: itemInfo.name,
           balance: itemBalance,
           price: tmpPrice,
-          logo: itemInfo.image ? 'https://ethplorer.io'+itemInfo.image : '',
+          logo: itemInfo.image ? 'https://ethplorer.io' + itemInfo.image : '',
           rate: itemInfo.price.rate,
           dif: itemInfo.price.diff,
         });
@@ -43,7 +44,7 @@ const CreateTokens = () => {
     })
   }
 
-  return {tokens: tokensForRender, totalSum: totalSum};
+  return {tokens: tokensForRender, totalSum};
 };
 
 const TokensPage = () => {
