--- src/components/NavBar/NavBar.css
@@ -1,11 +1,0 @@
-.ant-page-header {
-    background: #9fc4f1;
-}
-
-.ant-page-header-heading-sub-title {
-    font-size: 20px;
-}
-
-.navbar-slogan {
-    font-size: 3.8rem;
-}
