+++ src/components/NavBar/NavBar.scss
@@ -1,0 +1,9 @@
+@import "src/styles/variables";
+
+.ant-page-header {
+    background: $color-bg-blue;
+}
+
+.navbar-slogan {
+    font-size: 3.8rem;
+}
