+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { ConfigProvider } from 'antd';
 import 'antd/dist/antd.css';
 import App from './App';
-import './styles/index.css';
+import './styles/index.scss';
 
 
 ReactDOM.render(
@@ -14,45 +14,3 @@ ReactDOM.render(
   </React.StrictMode>,
   document.getElementById('root')
 );
-
-
-const DESKTOP_WIDTH = 1440;
-const DESKTOP_HEIGHT = 1024;
-const MOBILE_WIDTH = 720;
-const MOBILE_HEIGHT = 1280;
-
-const SIZE = {
-  MOBILE: [MOBILE_WIDTH, MOBILE_HEIGHT],
-};
-
-
-function fit() {
-
-  let  currentWidth = window.innerWidth;
-  let  currentHeight = window.innerHeight;
-
-
-  const initScale = 10;
-  let currentScale = 1;
-
-  const [aspectWidth, aspectHeight] = SIZE.MOBILE ;
-  if (currentWidth < aspectWidth) {
-    currentScale = currentWidth / aspectWidth;
-  }
-  if (currentHeight < aspectHeight) {
-    currentScale = Math.min(currentScale, currentHeight / aspectHeight);
-  }
-
-  currentScale = Math.max(currentScale, 0.5);
-
-  document.documentElement.style.fontSize = `${currentScale * initScale}px`;
-}
-
-function initScale() {
-  window.addEventListener("resize", fit);
-  fit();
-}
-
-document.addEventListener("DOMContentLoaded", () => {
-  initScale();
-})
