--- src/styles/mixins.scss
@@ -1,4 +1,0 @@
-@mixin square($size) {
-    width: $size;
-    height: $size;
-}
