+++ src/styles/variables.scss
@@ -1,2 +1,5 @@
-$color-grey: #fefefe;
-$color-bg-blue: #1a1a6e;
+$color-bg-blue: #9fc4f1;
+$color-green: green;
+$color-red: red;
+$color-white: white;
+$color-light-dark: rgba(0, 0, 0, 0.45);
