+++ src/utils/formatMoney.js
@@ -1,7 +1,7 @@
 
 
 export const formatMoney = (number, decimal) => {
-  let strNumber = String(number);
+  const strNumber = String(number);
   const idxDel = strNumber.indexOf('.');
   const mod = strNumber.slice(idxDel);
   const count = mod.search(/[1-9]/);
@@ -10,9 +10,9 @@ export const formatMoney = (number, decimal) => {
     decimal = count;
   }
 
-  number = Number(number);
+  const typeNumber = Number(number);
 
-  const rounded = Number(number.toFixed(decimal));
+  const rounded = Number(typeNumber.toFixed(decimal));
   if (rounded < 10000) {
     return rounded;
   }
